refactor: ProcessDefinitionMessageSubscriptionStatistics returns a page instead of a endCursor
